Added edit profile ability. Logs user out on password change though.

+++ QuizZing/src/main/java/com/UndefinedParameter/app/core/NewsManager.java
@@ -23,7 +23,7 @@ public class NewsManager {
 	}
 	
 	
-	public List<NewsArticle> getRecentNewsByUser(int userId) {
+	public List<NewsArticle> getRecentNewsByUser(long userId) {
 		//TODO: Select news by user id, format and return
 		return newsDAO.findAllNews();
 	}
